+++ CHANGES.md
@@ -1,3 +1,8 @@
+In next release ...
+
+- Use `bigint` everywhere as a type instead of `BigInt`.
+
+
 1.4.0 (2023-11-10)
 ------------------